requirements.txt
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.3
PyQt6-sip>=13.4.1
pynput>=1.7.6
sounddevice>=0.4.6
numpy
openai>=1.3.0
requests>=2.31.0
pyperclip>=1.8.2
pyautogui>=0.9.54
pywin32>=306
emoji>=2.2.0
pyinstaller>=6.3.0
google-generativeai>=0.3.2
